Implement warnings for number literals - `#` in numbers is deprecated - exponent markers that are not `e` are deprecated - removed support for single precision floats from memory/runtime - discovered plenty of bugs

+++ cli/src/main.rs
@@ -124,7 +124,7 @@ fn run() -> Result<()> {
     run::run(sources, config)?;
 
     let mut bytecode = BytecodeChunk::new("Test Chunk");
-    let num = bytecode.single(1.2);
+    let num = bytecode.double(1.2);
     let num = bytecode.number(num);
     bytecode.location(11);
     bytecode.write(OpCode::LoadConstant);
